API-login2-main/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
flask-migrate
pymysql
pyjwt